Remove StyledPropertyBase class.
Originally `StyledPropertyBase` was the base class for `StyledProperty` and `AttachedProperty` however #1499 made `AttachedProperty` derive directly from `StyledProperty` meaning that there is no longer any need for a separate `StyledPropertyBase` class.
