+++ src/Avalonia.Base/PropertyStore/ImmediateValueEntry.cs
@@ -9,7 +9,7 @@ namespace Avalonia.PropertyStore
 
         public ImmediateValueEntry(
             ImmediateValueFrame owner,
-            StyledPropertyBase<T> property, 
+            StyledProperty<T> property, 
             T value)
         {
             _owner = owner;
@@ -17,7 +17,7 @@ namespace Avalonia.PropertyStore
             Property = property;
         }
 
-        public StyledPropertyBase<T> Property { get; }
+        public StyledProperty<T> Property { get; }
         public bool HasValue => true;
         AvaloniaProperty IValueEntry.Property => Property;
